requirements.txt
certifi==2023.7.22
datadog-api-client==2.17.0
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
